// src/employees/enums/ResponseMessage.enum.ts
export enum ResponseMessage {
  RECORD_CREATED = 'Record created successfully',
  RECORD_UPDATED = 'Record updated successfully',
  RECORD_DELETED = 'Record deleted successfully',
  RECORDS_RETRIEVED = 'Records retrived successfully',
  INTERNAL_SERVER_ERROR = 'Something went wrong',
  BAD_REQUEST = 'Bad Request, No record present with this ID',
}